+++ gears/materials.py
@@ -7,6 +7,7 @@ from pbge import Singleton
 
 RT_REPAIR,RT_MEDICINE,RT_BIOTECHNOLOGY = list(range(3))
 
+
 class Material( Singleton ):
     # This base class is just used so all other materials count as subclasses
     # of "Material". I don't think this is an elegant way to handle things,
@@ -20,6 +21,8 @@ class Material( Singleton ):
     repair_type = RT_REPAIR
     repair_cost = 1
     BONUS_ACTION_COST_MOD = 0
+    DESTROYED_SOUND_FX = "hq-explosion-6288.ogg"
+
 
 class Metal( Material ):
     desig = "M"
@@ -29,6 +32,7 @@ class Metal( Material ):
     cost_scale = 10
     repair_type = RT_REPAIR
 
+
 class Advanced( Material ):
     desig = "A"
     name = "Advanced"
@@ -36,6 +40,7 @@ class Advanced( Material ):
     damage_scale = 1.0
     cost_scale = 20
     repair_type = RT_REPAIR
+
 
 class Ceramic( Material ):
     desig = "C"
@@ -45,6 +50,7 @@ class Ceramic( Material ):
     cost_scale = 15
     repair_type = RT_REPAIR
 
+
 class Meat( Material ):
     desig = "meat"
     name = "Meat"
@@ -53,6 +59,8 @@ class Meat( Material ):
     cost_scale = 7
     repair_type = RT_MEDICINE
     repair_cost = 2
+    DESTROYED_SOUND_FX = "cqccreaturedie7.ogg"
+
 
 class Biotech( Material ):
     desig = "B"
@@ -63,11 +71,14 @@ class Biotech( Material ):
     repair_type = RT_BIOTECHNOLOGY
     repair_cost = 2
     BONUS_ACTION_COST_MOD = 2
+    DESTROYED_SOUND_FX = "cqccreaturedie7.ogg"
+
 
 class DamageMat(Material):
     # Only used for rolling damage. Do not use this for a gear. I'll only warn you once.
     #  I never give a warning twice
     # Anything could happen like the rollin' of the dice - now!
     damage_scale = 0.4
+
 
 MECHA_MATERIALS = (Metal,Advanced,Ceramic,Biotech)